EE-287: Separate out ImportResolver.
Create specific instance of `ModuleImportResolver` based on a
`protocol_version` value. This assumes the following:

- New FFI functions can be added to the current resolver (i.e. V1)
- FFI functions can be removed from V2 resolver, but the codepaths
  for removed functions have to be supported in the Runtime and in
  the Context.
- Protocol upgrade would require implementation of separate resolver,
  possibly a resolver which would fall back to previous version
  resolver for a set of functions that doesn't change.
