+++ execution-engine/engine/src/functions.rs
@@ -1,0 +1,18 @@
+pub const WRITE_FUNC_INDEX: usize = 0;
+pub const READ_FUNC_INDEX: usize = 1;
+pub const ADD_FUNC_INDEX: usize = 2;
+pub const NEW_FUNC_INDEX: usize = 3;
+pub const GET_READ_FUNC_INDEX: usize = 4;
+pub const SER_FN_FUNC_INDEX: usize = 5;
+pub const GET_FN_FUNC_INDEX: usize = 6;
+pub const LOAD_ARG_FUNC_INDEX: usize = 7;
+pub const GET_ARG_FUNC_INDEX: usize = 8;
+pub const RET_FUNC_INDEX: usize = 9;
+pub const GET_CALL_RESULT_FUNC_INDEX: usize = 10;
+pub const CALL_CONTRACT_FUNC_INDEX: usize = 11;
+pub const GET_UREF_FUNC_INDEX: usize = 12;
+pub const GAS_FUNC_INDEX: usize = 13;
+pub const HAS_UREF_FUNC_INDEX: usize = 14;
+pub const ADD_UREF_FUNC_INDEX: usize = 15;
+pub const STORE_FN_INDEX: usize = 16;
+pub const PROTOCOL_VERSION_FUNC_INDEX: usize = 17;
