+++ execution-engine/engine/src/resolvers/error.rs
@@ -1,0 +1,5 @@
+#[derive(Debug)]
+pub enum ResolverError {
+    UnknownProtocolVersion(u64),
+    NoImportedMemory,
+}
